refactor(PermissionUtil): 권한을 요청하기 위해 다이얼로그를 띄우는 기능을 SettingFragment.kt 파일에서 PermissionUtil.kt 파일로 옮김
